Extensive refactoring of our Query architecture.  Queries are now much faster. Moved relevancy calculations into the backends. Removed dead, unused and obsolete code.

+++ Tiles/TileHitCollection.cs
@@ -68,7 +68,7 @@ namespace Beagle.Tile {
 		Template head_template;
 		Template foot_template;
 
-		public float MaxScore {
+		public double MaxScore {
 			get {
 				if (all_hits.Count == 0)
 					return 0;
